requirements_local.txt
# 本地化语音处理依赖
openai-whisper>=20231117
pyttsx3>=2.90
argostranslate>=1.9.0
pydub>=0.25.1
moviepy>=1.0.3
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0
requests>=2.25.0
